requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
requests==2.31.0
scikit-learn==1.4.0
plotly==5.18.0
networkx==3.2.1
python-dotenv==1.0.1
matplotlib==3.8.3
seaborn==0.13.2 